refactor: extract AgentOrchestrator into agent/ sub-modules
Split the monolithic AgentOrchestrator.ts (699 lines) into focused
sub-modules under src/services/agent/:

- prompt-builder.ts: buildAgentPrompt + DEEP_RESEARCH_PROTOCOL constant
- process-manager.ts: spawn, stdout/stderr parsing, timeout, cleanup
- worktree.ts: git worktree create/remove helpers
- types.ts: shared types (AgentProcess, AgentLogEvent, ProcessManagerDeps)
- index.ts: facade re-exporting service, Layer, and all public API

spawnAgent reduced from ~300 lines to ~50 lines of orchestration.
All consumers updated to import from new path.
