+++ src/services/Heartbeat.ts
@@ -13,7 +13,7 @@ import { Context, Effect, Layer } from "effect"
 import cron from "node-cron"
 import { ConfigService } from "./Config.js"
 import { Kanban } from "./Kanban.js"
-import { AgentOrchestrator } from "./AgentOrchestrator.js"
+import { AgentOrchestrator } from "./agent/index.js"
 import { AppPersistence, type AppKanbanCard } from "./AppPersistence.js"
 import { Telegram } from "./Telegram.js"
 import { ClaudeMem } from "./ClaudeMem.js"
